requirements.txt
gradio==4.44.0
pinecone-client==3.0.0
openai==0.28.1
sentence-transformers>=2.5.0
tavily-python==0.3.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
numpy>=1.26.0
pandas>=2.1.0
asyncio-mqtt==0.16.1
pydantic==2.5.0
uvicorn==0.24.0
fastapi==0.104.1
huggingface-hub>=0.20.0